refactor BaseDiffractionVectors for n-dimensional methods
