+++ pyxem/signals/diffraction_vectors2d.py
@@ -18,24 +18,22 @@
 
 import numpy as np
 
-from hyperspy.signals import BaseSignal, Signal1D
 from hyperspy.api import markers
 
 import matplotlib.pyplot as plt
 from matplotlib.cm import get_cmap
 from scipy.spatial import distance_matrix
-from sklearn.cluster import DBSCAN
 
-from warnings import warn
+from pyxem.signals.diffraction_vectors import BaseDiffractionVectors
 
 from pyxem.signals import push_metadata_through
-from pyxem.signals import transfer_navigation_axes
-from pyxem.utils.vector_utils import calculate_norms, calculate_norms_ragged
-from pyxem.utils.vector_utils import get_npeaks
 from pyxem.utils.plot import generate_marker_inputs_from_peaks
 
 """
-Signal class for diffraction vectors.
+Signal class for two-dimensional diffraction vectors.
+
+Typically obtained from diffraction data recorded on a two-dimensional detector
+under the flat Ewald sphere approximation.
 
 There are two cases that are supported:
 
@@ -48,7 +46,7 @@ number of peaks.
 """
 
 
-class DiffractionVectors2D(BaseSignal):
+class DiffractionVectors2D(BaseDiffractionVectors):
     """Two-dimensional diffraction vectors in reciprocal Angstrom units.
 
     Attributes
@@ -56,9 +54,6 @@ class DiffractionVectors2D(BaseSignal):
     detector_coordinates : DetectorCoordinates2D
         Array of 2-vectors describing detector coordinates associated with each
         diffraction vector.
-    cartesian : np.array
-        Array of 3-vectors describing Cartesian coordinates associated with
-        each diffraction vector.
     hkls : np.array
         Array of Miller indices associated with each diffraction vector
         following indexation.
@@ -69,18 +64,36 @@ class DiffractionVectors2D(BaseSignal):
         self, args, kwargs = push_metadata_through(self, *args, **kwargs)
         super().__init__(*args, **kwargs)
         self.detector_coordinates = None
-        self.cartesian = None
         self.hkls = None
 
-    def plot_vectors(self, xlim=1.0, ylim=1.0,
-                     unique_vectors=None,
-                     distance_threshold=0.01,
-                     method='distance_comparison',
-                     min_samples=1,
-                     image_to_plot_on=None,
-                     image_cmap='gray',
-                     plot_label_colors=False,
-                     distance_threshold_all=0.005):  # pragma: no cover
+    def plot_vectors_on_signal(self, signal, *args, **kwargs):
+        """Plot the diffraction vectors on a signal.
+
+        Parameters
+        ----------
+        signal : ElectronDiffraction2D
+            The ElectronDiffraction2D signal object on which to plot the peaks.
+            This signal must have the same navigation dimensions as the peaks.
+        *args :
+            Arguments passed to signal.plot()
+        **kwargs :
+            Keyword arguments passed to signal.plot()
+        """
+        mmx, mmy = generate_marker_inputs_from_peaks(self)
+        signal.plot(*args, **kwargs)
+        for mx, my in zip(mmx, mmy):
+            m = markers.point(x=mx, y=my, color='red', marker='x')
+            signal.add_marker(m, plot_marker=True, permanent=False)
+
+    def plot_unique_vectors(self, xlim=1.0, ylim=1.0,
+                            unique_vectors=None,
+                            distance_threshold=0.01,
+                            method='distance_comparison',
+                            min_samples=1,
+                            image_to_plot_on=None,
+                            image_cmap='gray',
+                            plot_label_colors=False,
+                            distance_threshold_all=0.005):  # pragma: no cover
         """Plot the unique diffraction vectors.
 
         Parameters
@@ -154,8 +167,8 @@ class DiffractionVectors2D(BaseSignal):
             # Get all vectors from the clustering not considered noise
             cores = clusters.components_
             if cores.size == 0:
-                warn('No clusters were found. Check parameters, or '
-                     'use plot_label_colors=False.')
+                ValueError('No clusters were found. Check parameters, or '
+                           'use plot_label_colors=False.')
             else:
                 peaks = DiffractionVectors2D(cores)
                 peaks.axes_manager.set_signal_dimension(1)
@@ -197,210 +210,3 @@ class DiffractionVectors2D(BaseSignal):
         plt.tight_layout()
         plt.axis('off')
         return fig
-
-    def plot_vectors_on_signal(self, signal, *args, **kwargs):
-        """Plot the diffraction vectors on a signal.
-
-        Parameters
-        ----------
-        signal : ElectronDiffraction2D
-            The ElectronDiffraction2D signal object on which to plot the peaks.
-            This signal must have the same navigation dimensions as the peaks.
-        *args :
-            Arguments passed to signal.plot()
-        **kwargs :
-            Keyword arguments passed to signal.plot()
-        """
-        mmx, mmy = generate_marker_inputs_from_peaks(self)
-        signal.plot(*args, **kwargs)
-        for mx, my in zip(mmx, mmy):
-            m = markers.point(x=mx, y=my, color='red', marker='x')
-            signal.add_marker(m, plot_marker=True, permanent=False)
-
-    def get_magnitudes(self, *args, **kwargs):
-        """Calculate the magnitude of diffraction vectors.
-
-        Parameters
-        ----------
-        *args:
-            Arguments to be passed to map().
-        **kwargs:
-            Keyword arguments to map().
-
-        Returns
-        -------
-        magnitudes : BaseSignal
-            A signal with navigation dimensions as the original diffraction
-            vectors containging an array of gvector magnitudes at each
-            navigation position.
-
-        """
-        # If ragged the signal axes will not be defined
-        if len(self.axes_manager.signal_axes) == 0:
-            magnitudes = self.map(calculate_norms_ragged,
-                                  inplace=False,
-                                  *args, **kwargs)
-        # Otherwise easier to calculate.
-        else:
-            magnitudes = BaseSignal(calculate_norms(self))
-            magnitudes.axes_manager.set_signal_dimension(0)
-
-        return magnitudes
-
-    def get_magnitude_histogram(self, bins, *args, **kwargs):
-        """Obtain a histogram of gvector magnitudes.
-
-        Parameters
-        ----------
-        bins : numpy array
-            The bins to be used to generate the histogram.
-        *args:
-            Arguments to get_magnitudes().
-        **kwargs:
-            Keyword arguments to get_magnitudes().
-
-        Returns
-        -------
-        ghis : Signal1D
-            Histogram of gvector magnitudes.
-
-        """
-        gmags = self.get_magnitudes(*args, **kwargs)
-
-        if len(self.axes_manager.signal_axes) == 0:
-            glist = []
-            for i in gmags._iterate_signal():
-                for j in np.arange(len(i[0])):
-                    glist.append(i[0][j])
-            gs = np.asarray(glist)
-            gsig = Signal1D(gs)
-            ghis = gsig.get_histogram(bins=bins)
-
-        else:
-            ghis = gmags.get_histogram(bins=bins)
-
-        ghis.axes_manager.signal_axes[0].name = 'k'
-        ghis.axes_manager.signal_axes[0].units = '$A^{-1}$'
-
-        return ghis
-
-    def get_unique_vectors(self, distance_threshold=0.01,
-                           method='distance_comparison', min_samples=1,
-                           return_clusters=False):
-        """Returns diffraction vectors considered unique by:
-        strict comparison, distance comparison with a specified
-        threshold, or by clustering using DBSCAN [1].
-
-        Parameters
-        ----------
-        distance_threshold : float
-            The minimum distance between diffraction vectors for them to
-            be considered unique diffraction vectors. If
-            distance_threshold==0, the unique vectors will be determined
-            by strict comparison.
-        method : str
-            The method to use to determine unique vectors. Valid methods
-            are 'strict', 'distance_comparison' and 'DBSCAN'.
-            'strict' returns all vectors that are strictly unique and
-            corresponds to distance_threshold=0.
-            'distance_comparison' checks the distance between vectors to
-            determine if some should belong to the same unique vector,
-            and if so, the unique vector is iteratively updated to the
-            average value.
-            'DBSCAN' relies on the DBSCAN [1] clustering algorithm, and
-            uses the Eucledian distance metric.
-        min_samples : int, optional
-            The minimum number of not strictly identical vectors within
-            one cluster for the cluster to be considered a core sample,
-            i.e. to not be considered noise. Only used for method='DBSCAN'.
-        return_clusters : bool, optional
-            If True (False is default), the DBSCAN clustering result is
-            returned. Only used for method='DBSCAN'.
-
-        References
-        ----------
-        [1] https://scikit-learn.org/stable/modules/generated/sklearn.
-            cluster.DBSCAN.html
-
-        Returns
-        -------
-        unique_peaks : DiffractionVectors2D
-            The unique diffraction vectors.
-        clusters : DBSCAN
-            The results from the clustering, given as class DBSCAN.
-            Only returned if method='DBSCAN' and return_clusters=True.
-        """
-        # Flatten the array of peaks to reach dimension (n, 2), where n
-        # is the number of peaks.
-        peaks_all = np.concatenate([
-            peaks.ravel() for peaks in self.data.flat]).reshape(-1, 2)
-
-        # A distance_threshold of 0 implies a strict comparison. So in that
-        # case, a warning is raised unless the specified method is 'strict'.
-        if distance_threshold == 0:
-            if method is not 'strict':
-                warn(message='distance_threshold=0 was given, and therefore ' +
-                     'a strict comparison is used, even though the ' +
-                     'specified method was ' + method + '.')
-                method = 'strict'
-
-        if method == 'strict':
-            unique_peaks = np.unique(peaks_all, axis=0)
-
-        elif method == 'distance_comparison':
-            unique_vectors, unique_counts = np.unique(
-                peaks_all, axis=0, return_counts=True)
-
-            unique_peaks = np.array([[0, 0]])
-            unique_peaks_counts = np.array([0])
-
-            while unique_vectors.shape[0] > 0:
-                unique_vector = unique_vectors[0]
-                distances = distance_matrix(
-                    np.array([unique_vector]), unique_vectors)
-                indices = np.where(distances < distance_threshold)[1]
-
-                new_count = indices.size
-                new_unique_peak = np.array([np.average(
-                    unique_vectors[indices], weights=unique_counts[indices],
-                    axis=0)])
-
-                unique_peaks = np.append(unique_peaks, new_unique_peak,
-                                         axis=0)
-
-                unique_peaks_counts = np.append(unique_peaks_counts,
-                                                new_count)
-                unique_vectors = np.delete(unique_vectors, indices, axis=0)
-                unique_counts = np.delete(unique_counts, indices, axis=0)
-            unique_peaks = np.delete(unique_peaks, [0], axis=0)
-
-        elif method == 'DBSCAN':
-            # All peaks are clustered by DBSCAN so that peaks within
-            # one cluster are separated by distance_threshold or less.
-            unique_vectors, unique_vectors_counts = np.unique(
-                peaks_all, axis=0, return_counts=True)
-            clusters = DBSCAN(
-                eps=distance_threshold, min_samples=min_samples,
-                metric='euclidean').fit(
-                unique_vectors, sample_weight=unique_vectors_counts)
-            unique_labels, unique_labels_count = np.unique(
-                clusters.labels_, return_counts=True)
-            unique_peaks = np.zeros((unique_labels.max() + 1, 2))
-            # For each cluster, a center of mass is calculated based
-            # on all the peaks within the cluster, and the center of
-            # mass is taken as the final unique vector position.
-            for n in np.arange(unique_labels.max() + 1):
-                peaks_n_temp = unique_vectors[clusters.labels_ == n]
-                peaks_n_counts_temp = unique_vectors_counts[
-                    clusters.labels_ == n]
-                unique_peaks[n] = np.average(
-                    peaks_n_temp, weights=peaks_n_counts_temp,
-                    axis=0)
-        # Manipulate into DiffractionVectors2D class
-        if unique_peaks.size > 0:
-            unique_peaks = DiffractionVectors2D(unique_peaks)
-            unique_peaks.axes_manager.set_signal_dimension(1)
-        if return_clusters and method == 'DBSCAN':
-            return unique_peaks, clusters
-        else:
-            return unique_peaks
